Added physics support
The physics support at this point just allows the contact with the
first pin by the ball to be detected. When detected, some pins are
removed as an example.

If the code to remove the pins and ball is removed from the didBeginContact,
you can see how the ball affects the first pin (it's moved out of the way).
When I turned on physics support for all of the pins, they moved away from
each other which was interesting

Finally, the part that was missing to get the physics working was adding the
gravity setting. I guess that's why the physics-enabled objects just dropped
off the scene.